# openapi.yaml
openapi: 3.1.0
info:
  title: Books & Finance API
  version: 1.0.0
  description: |
    Personal finance and books management REST API.

    **Base URL:** `http://localhost:8000/api/v1`

    **Authentication:** Bearer JWT token. Obtain via `POST /auth/login`.
    Tokens expire in **7 days**.

    **Default credentials:** `BTC007` / `password123`
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    search:
      name: search
      in: query
      schema: { type: string }
    from:
      name: from
      in: query
      schema: { type: string, format: date }
    to:
      name: to
      in: query
      schema: { type: string, format: date }
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  maxLength: 50
                  example: john_doe
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  minLength: 6
                  example: secret123
      responses:
        '201':
          description: Account created
          content:
            application/json:
              example:
                success: true
                data:
                  token: eyJhbGci...
                  user: { id: 2, username: john_doe, email: john@example.com }
                message: Account created successfully
        '409':
          description: Email or username already exists
  /auth/login:
    post:
      summary: Login and receive a JWT token
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                email:
                  type: string
                  format: email
                  example: btc007@books.local
                username:
                  type: string
                  example: BTC007
                password:
                  type: string
                  example: password123
            examples:
              byEmail:
                summary: Login by email
                value: { email: btc007@books.local, password: password123 }
              byUsername:
                summary: Login by username
                value: { username: BTC007, password: password123 }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              example:
                success: true
                data:
                  token: eyJhbGci...
                  expiresIn: 7d
                  user: { id: 1, username: BTC007, email: btc007@books.local }
                message: Logged in successfully
        '401':
          description: Invalid credentials
...
  /accounts:
    get:
      summary: List accounts
      tags: [Accounts]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/search'
      responses:
        '200':
          description: List of accounts
    post:
      summary: Create account
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type]
              properties:
                name: { type: string, example: Main Wallet }
                type: { type: string, example: bank }
                balance: { type: number, example: 5000 }
                notes: { type: string }
      responses:
        '201':
          description: Account created
  /accounts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      summary: Get account
      tags: [Accounts]
      responses:
        '200': { description: Account found }
        '404': { description: Not found }
    patch:
      summary: Update account
      tags: [Accounts]
      requestBody:
        content:
          application/json:
            example: { name: Updated Name, balance: 7500 }
      responses:
        '200': { description: Updated }
    delete:
      summary: Delete account
      tags: [Accounts]
      responses:
        '204': { description: Deleted }
...
  /expenses:
    get:
      summary: List expenses
      tags: [Expenses]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
      responses:
        '200': { description: Expenses list }
    post:
      summary: Add expense
      tags: [Expenses]
      requestBody:
        required: true
        content:
          application/json:
            example:
              category: Food
              amount: 45.00
              date: "2024-01-15"
              description: Dinner
      responses:
        '201': { description: Expense created }
...
  /financial-plans:
    get:
      summary: List financial plans
      tags: [Financial Plans]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema: { type: string, enum: [draft, active, completed, archived] }
      responses:
        '200': { description: Plans list }
    post:
      summary: Create financial plan
      tags: [Financial Plans]
      requestBody:
        required: true
        content:
          application/json:
            example:
              title: Q1 2024 Budget Plan
              description: Quarterly budget allocation
              start_date: "2024-01-01"
              end_date: "2024-03-31"
              status: active
      responses:
        '201': { description: Plan created }
...