# routes/auth.js
const express = require('express');
const router = express.Router();
const jwt = require('jsonwebtoken');
const bcrypt = require('bcryptjs');
const { z } = require('zod');

const db = require('../db/connection');
const { sendSuccess } = require('../utils/response');
const { authLimiter } = require('../middleware/rateLimiter');
const validate = require('../middleware/zodValidate');
const AppError = require('../utils/AppError');
const asyncHandler = require('../utils/asyncHandler');

// ── Zod Schemas ───────────────────────────────────────────────────────────────
const registerSchema = z.object({
  body: z.object({
    username: z.string().min(1, 'Username is required').max(50),
    email: z.string().email('Valid email required'),
    password: z.string().min(6, 'Password must be at least 6 characters'),
  })
});

const loginSchema = z.object({
  body: z.object({
    email: z.string().email('Valid email required').optional(),
    username: z.string().min(1).max(50).optional(),
    password: z.string().min(1, 'Password is required'),
  }).refine(data => data.email || data.username, {
    message: 'Provide email or username',
    path: ['email']
  })
});

// ── Utilities ────────────────────────────────────────────────────────────────
const TokenService = require('../utils/tokenService');

// ── POST /auth/register ───────────────────────────────────────────────────────
router.post('/register', authLimiter, validate(registerSchema), asyncHandler(async (req, res) => {
  const { username, email, password } = req.body;

  const existingEmail = await db.prepare('SELECT id FROM users WHERE email = ?').get(email);
  if (existingEmail) throw new AppError('Email already registered', 409, 'CONFLICT');

  const existingUsername = await db.prepare('SELECT id FROM users WHERE username = ?').get(username);
  if (existingUsername) throw new AppError('Username already taken', 409, 'CONFLICT');

  const hash = bcrypt.hashSync(password, 10);
  const now = new Date().toISOString();

  const info = await db.prepare(
    'INSERT INTO users (username, email, password_hash, role, created_at, updated_at) VALUES (?, ?, ?, ?, ?, ?)'
  ).run(username, email, hash, 'user', now, now);

  const user = await db.prepare('SELECT id, username, email, role, created_at FROM users WHERE id = ?').get(info.lastInsertRowid);

  const { accessToken, refreshToken } = await TokenService.issueEnhancedTokens(user);

  return sendSuccess(res, { accessToken, refreshToken, user }, 'Account created successfully', 201);
}));

// ── POST /auth/login ──────────────────────────────────────────────────────────
router.post('/login', authLimiter, validate(loginSchema), asyncHandler(async (req, res) => {
  const { email, username, password } = req.body;

  const user = email
    ? await db.prepare('SELECT * FROM users WHERE email = ?').get(email)
    : await db.prepare('SELECT * FROM users WHERE username = ?').get(username);

  if (!user) throw new AppError('Invalid credentials', 401, 'UNAUTHORIZED');

  const isMatch = bcrypt.compareSync(password, user.password_hash);
  if (!isMatch) throw new AppError('Invalid credentials', 401, 'UNAUTHORIZED');

  const { accessToken, refreshToken } = await TokenService.issueEnhancedTokens(user);

  const safeUser = { id: user.id, username: user.username, email: user.email, role: user.role, created_at: user.created_at };
  return sendSuccess(res, { accessToken, refreshToken, user: safeUser }, 'Logged in successfully');
}));

// ── POST /auth/refresh ───────────────────────────────────────────────────────
router.post('/refresh', authLimiter, asyncHandler(async (req, res) => {
  const { refreshToken } = req.body;
  if (!refreshToken) throw new AppError('Refresh token required', 400, 'BAD_REQUEST');

  try {
    const tokens = await TokenService.rotateRefreshToken(refreshToken);
    return sendSuccess(res, tokens, 'Token refreshed successfully');
  } catch (err) {
    throw new AppError(err.message || 'Invalid refresh token', 401, 'UNAUTHORIZED');
  }
}));

// ── POST /auth/logout ────────────────────────────────────────────────────────
router.post('/logout', asyncHandler(async (req, res) => {
  const { refreshToken } = req.body;
  if (refreshToken) {
    await TokenService.revokeToken(refreshToken);
  }
  return sendSuccess(res, null, 'Logged out successfully');
}));

// ── POST /auth/logout-all ────────────────────────────────────────────────────
router.post('/logout-all', asyncHandler(async (req, res) => {
  const { refreshToken } = req.body;
  if (refreshToken) {
    // Determine userId from format b64Id.rand
    if (refreshToken.includes('.')) {
      const [b64Id, ] = refreshToken.split('.');
      const userId = Buffer.from(b64Id, 'base64').toString('utf8');
      if (userId) {
        await TokenService.revokeAllUserTokens(userId);
      }
    }
  }
  return sendSuccess(res, null, 'Logged out of all sessions successfully');
}));

module.exports = router;